seperation of platforms
